get_model_s3(): - rm from: gamechangerml/src/utilities/utils.py - add to new file: gamechangerml/data_transfer/s3_download.py
