+++ gamechangerml/data_transfer/__init__.py
@@ -1,1 +1,2 @@
 from .delete_local_corpus import delete_local_corpus
+from .s3_download import get_model_s3
