+++ gamechangerml/src/utilities/utils.py
@@ -14,45 +14,6 @@ from gamechangerml.utils import configure_logger
 
 logger = logging.getLogger("gamechanger")
 
-def get_model_s3(
-    filename, s3_model_dir, download_dir="", bucket=None, logger=None
-):
-    """Download a model from S3.
-
-    Args:
-        filename (str): File name of the model to download.
-        s3_model_dir (str): Path to S3 directory which contains the model to 
-            download.
-        download_dir (str, optional): Path to local directory to put downloaded
-            files. Defaults to "".
-        bucket (boto3.resources.factory.s3.Bucket or None, optional): Bucket to
-            upload to. If None, uses S3Service.connect_to_bucket(). Default is 
-            None.
-        logger (logging.Logger or None, optional): If None, uses 
-            configure_logger(). Default is None.
-
-    Returns:
-        list of str: Paths to locally downloaded files.
-    """
-    if logger is None:
-        logger = configure_logger()
-
-    files = []
-    model_path = join(s3_model_dir, filename)
-
-    try:
-        for obj in bucket.objects.filter(Prefix=model_path):
-            if obj.size != 0:
-                logger.info(f"Downloading {obj.key}")
-                bucket.download_file(
-                    obj.key, join(download_dir, obj.key.split("/")[-1])
-                )
-                files.append(join(download_dir, obj.key.split("/")[-1]))
-    except Exception:
-        logger.exception(f"Failed to download S3 model at {model_path}.")
-
-    return files
-
 def verify_model_name(model_dir, filePrefix):
     count = 0
